+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Marek Kurańda](https://github.com/mjkuranda): GraphQL management and `resolver` and `schema`.
 - [Marek Kurańda](https://github.com/mjkuranda): `PropertyService` and all instances of services and repositories.
 - [Marek Kurańda](https://github.com/mjkuranda): Connection to the database.
+- [Marek Kurańda](https://github.com/mjkuranda): Unit tests for services.
 
 ## [0.3.0] - 2025-03-07
 ### Added